Move all non-click functions in main to a file where they're testable
